@@ -1,13 +1,15 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Backend.Fx.DependencyInjection;
 using Backend.Fx.ExecutionPipeline;
-using Backend.Fx.Extensions;
+using Backend.Fx.Features;
 using Backend.Fx.Logging;
 using Backend.Fx.Util;
+using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 
 namespace Backend.Fx
@@ -15,6 +17,7 @@ namespace Backend.Fx
     /// <summary>
     /// The root object of the whole backend fx application framework
     /// </summary>
+    [PublicAPI]
     public interface IBackendFxApplication : IDisposable
     {
         /// <summary>
@@ -45,15 +48,31 @@ namespace Backend.Fx
         /// <returns></returns>
         Task BootAsync(CancellationToken cancellationToken = default);
 
-        TBackendFxApplicationDecorator As<TBackendFxApplicationDecorator>() where TBackendFxApplicationDecorator : BackendFxApplicationExtension;
+        /// <summary>
+        /// Enables an optional feature. Must be done before calling <see cref="BootAsync"/>.
+        /// </summary>
+        /// <param name="feature"></param>
+        void EnableFeature(Feature feature);
+
+        void RequireDependantFeature<TFeature>() where TFeature : Feature;
     }
 
 
     public class BackendFxApplication : IBackendFxApplication
     {
         private static readonly ILogger Logger = Log.Create<BackendFxApplication>();
-        private readonly ManualResetEventSlim _isBooted = new ManualResetEventSlim(false);
+        private readonly ManualResetEventSlim _isBooted = new(false);
+        private readonly List<Feature> _features = new();
 
+        /// <summary>
+        /// Initializes the application's runtime instance
+        /// </summary>
+        /// <param name="compositionRoot">The composition root of the dependency injection framework</param>
+        /// <param name="assemblies"></param>
+        public BackendFxApplication(ICompositionRoot compositionRoot, params Assembly[] assemblies)
+        : this(compositionRoot, new ExceptionLogger(Logger), assemblies)
+        { }
+        
         /// <summary>
         /// Initializes the application's runtime instance
         /// </summary>
@@ -62,7 +81,7 @@ namespace Backend.Fx
         /// <param name="assemblies"></param>
         public BackendFxApplication(ICompositionRoot compositionRoot, IExceptionLogger exceptionLogger, params Assembly[] assemblies)
         {
-            assemblies = assemblies ?? Array.Empty<Assembly>();
+            assemblies ??= Array.Empty<Assembly>();
             
             Logger.LogInformation(
                 "Initializing application with {CompositionRoot} providing services from [{Assemblies}]",
@@ -73,7 +92,7 @@ namespace Backend.Fx
             
             Invoker = new ExceptionLoggingInvoker(exceptionLogger, invoker);
 
-            CompositionRoot = compositionRoot;
+            CompositionRoot = new LogRegistrationsDecorator(compositionRoot);
             ExceptionLogger = exceptionLogger;
             Assemblies = assemblies;
             CompositionRoot.RegisterModules(new ExecutionPipelineModule(withFrozenClockDuringExecution: true));
@@ -87,20 +106,39 @@ namespace Backend.Fx
 
         public IExceptionLogger ExceptionLogger { get; }
         
-        public bool IsMultiTenancyApplication { get; set; }
+        public virtual void EnableFeature(Feature feature)
+        {
+            if (_isBooted.IsSet)
+            {
+                throw new InvalidOperationException("Features must be enabled before booting the application");
+            }
+            
+            feature.Enable(this);
+            _features.Add(feature);
+        }
 
-        public Task BootAsync(CancellationToken cancellationToken = default)
+        public void RequireDependantFeature<TFeature>() where TFeature : Feature
+        {
+            if (!_features.OfType<TFeature>().Any())
+            {
+                throw new InvalidOperationException($"This feature requires the {typeof(TFeature).Name} to be enabled first");
+            }
+        }
+
+        public async Task BootAsync(CancellationToken cancellationToken = default)
         {
             Logger.LogInformation("Booting application");
             CompositionRoot.Verify();
-            _isBooted.Set();
-            return Task.CompletedTask;
-        }
 
-        public virtual TBackendFxApplicationDecorator As<TBackendFxApplicationDecorator>()
-            where TBackendFxApplicationDecorator : BackendFxApplicationExtension
-        {
-            return null;
+            foreach (Feature feature in _features)
+            {
+                if (feature is IBootableFeature bootableFeature)
+                {
+                    await bootableFeature.BootAsync(this, cancellationToken).ConfigureAwait(false);
+                }
+            }
+
+            _isBooted.Set();
         }
 
         public bool WaitForBoot(int timeoutMilliSeconds = int.MaxValue, CancellationToken cancellationToken = default)
@@ -111,8 +149,11 @@ namespace Backend.Fx
         public void Dispose()
         {
             Logger.LogInformation("Application shut down initialized");
+            foreach (var feature in _features)
+            {
+                feature.Dispose();
+            }
             CompositionRoot?.Dispose();
-            GC.SuppressFinalize(this);
         }
     }
 }
