@@ -94,8 +94,8 @@ namespace Backend.Fx.Util
         {
             private bool _done;
             public Exception InnerException { private get; set; }
-            private readonly AutoResetEvent _workItemsWaiting = new AutoResetEvent(false);
-            private readonly Queue<Tuple<SendOrPostCallback, object>> _items = new Queue<Tuple<SendOrPostCallback, object>>();
+            private readonly AutoResetEvent _workItemsWaiting = new(false);
+            private readonly Queue<Tuple<SendOrPostCallback, object>> _items = new();
 
             public override void Send(SendOrPostCallback d, object state)
             {
