@@ -4,12 +4,27 @@ using JetBrains.Annotations;
 namespace Backend.Fx.ExecutionPipeline
 {
     [PublicAPI]
-    public sealed class SystemIdentity : IIdentity
+    public struct SystemIdentity : IIdentity
     {
         public string Name => "SYSTEM";
 
-        public string AuthenticationType => "system internal";
+        public string AuthenticationType => "Internal";
 
         public bool IsAuthenticated => true;
+
+        public override bool Equals(object obj)
+        {
+            return obj is SystemIdentity;
+        }
+
+        public override int GetHashCode()
+        {
+            return 542451621;
+        }
+
+        public bool Equals(SystemIdentity other)
+        {
+            return true;
+        }
     }
 }
