@@ -1,23 +1,25 @@
-using System.Reflection;
 using Backend.Fx.Domain;
 using Backend.Fx.Exceptions;
+using Backend.Fx.Features.Persistence;
 using JetBrains.Annotations;
 
 namespace Backend.Fx.Features.Authorization
 {
-    public static class AuthorizationFeature
+    /// <summary>
+    /// The feature "Authorization" obligates you the implementation of an <see cref="IAuthorizationPolicy{TAggregateRoot, TId}"/>
+    /// for every <see cref="IAggregateRoot{TId}"/>. Instances of these policy classes are applied to the repositories, so
+    /// that on every read or write operation on it, the policy is automatically enforced. Denied reads won't fail but
+    /// just appear invisible, while a denied write throws a <see cref="ForbiddenException"/>.
+    /// While implementing policies, you can start by deriving from <see cref="DenyAll{TAggregateRoot, TId}"/> or
+    /// <see cref="AllowAll{TAggregateRoot, TId}"/>.
+    /// </summary>
+    [PublicAPI]
+    public class AuthorizationFeature : Feature
     {
-        /// <summary>
-        /// The feature "Authorization" obligates you the implementation of an <see cref="IAuthorizationPolicy{TAggregateRoot}"/>
-        /// for every <see cref="AggregateRoot"/>. Instances of these policy classes are applied to the repositories, so
-        /// that on every read or write operation on it, the policy is automatically enforced. Denied reads won't fail but
-        /// just appear invisible, while a denied write throws a <see cref="ForbiddenException"/>.
-        /// While implementing policies, you can start by deriving from <see cref="DenyAll{TAggregateRoot}"/> or
-        /// <see cref="AllowAll{TAggregateRoot}"/>.
-        /// </summary>
-        /// <param name="application"></param>
-        [PublicAPI]
-        public static void EnableAuthorization(this IBackendFxApplication application)
-            => application.CompositionRoot.RegisterModules(new AuthorizationModule(application.Assemblies));
+        public override void Enable(IBackendFxApplication application)
+        {
+            application.RequireDependantFeature<PersistenceFeature>();
+            application.CompositionRoot.RegisterModules(new AuthorizationModule(application.Assemblies));
+        }
     }
 }
