@@ -3,6 +3,7 @@ using System.Linq;
 using System.Reflection;
 using Backend.Fx.DependencyInjection;
 using Backend.Fx.Domain;
+using Backend.Fx.Features.Persistence;
 using Backend.Fx.Logging;
 using Backend.Fx.Util;
 using Microsoft.Extensions.DependencyInjection;
@@ -31,44 +32,54 @@ namespace Backend.Fx.Features.Authorization
         private static void RegisterAuthorizingDecorators(ICompositionRoot compositionRoot)
         {
             Logger.LogDebug("Registering authorization decorators");
-            compositionRoot.RegisterDecorator(ServiceDescriptor.Scoped(typeof(IAsyncAggregateQueryable<>), typeof(AuthorizingQueryable<>)));
-            compositionRoot.RegisterDecorator(ServiceDescriptor.Scoped(typeof(IRepository<>), typeof(AuthorizingRepository<>)));
+            compositionRoot.RegisterDecorator(ServiceDescriptor.Scoped(typeof(IAggregateQueryable<,>),
+                typeof(AuthorizingQueryable<,>)));
+            compositionRoot.RegisterDecorator(ServiceDescriptor.Scoped(typeof(IRepository<,>),
+                typeof(AuthorizingRepository<,>)));
         }
-        
+
         private void RegisterAuthorizationPolicies(ICompositionRoot compositionRoot)
         {
             Logger.LogDebug("Registering authorization services from {Assemblies}", _assembliesForLogging);
 
-            var aggregateRootTypes = _assemblies.GetImplementingTypes(typeof(AggregateRoot)).ToArray();
+            var aggregateRootTypes = _assemblies
+                .SelectMany(ass => ass.GetTypes())
+                .Where(t => t.IsImplementationOfOpenGenericInterface(typeof(IAggregateRoot<>)))
+                .ToArray();
+
             foreach (var aggregateRootType in aggregateRootTypes)
             {
-                var aggregateAuthorizationTypes = _assemblies
-                    .GetImplementingTypes(typeof(IAuthorizationPolicy<>).MakeGenericType(aggregateRootType))
+                var idType = aggregateRootType
+                    .GetInterfaces()
+                    .Where(inf => inf.IsGenericType)
+                    .Single(inf => inf.GetGenericTypeDefinition() == typeof(IAggregateRoot<>))
+                    .GenericTypeArguments[0];
+
+                var authorizationPolicyInterfaceType =
+                    typeof(IAuthorizationPolicy<,>).MakeGenericType(aggregateRootType, idType);
+                var authorizationPolicyTypes = _assemblies
+                    .GetImplementingTypes(authorizationPolicyInterfaceType)
                     .ToArray();
 
-                foreach (Type aggregateAuthorizationType in aggregateAuthorizationTypes)
+                if (authorizationPolicyTypes.Length == 0)
                 {
-                    var serviceTypes = aggregateAuthorizationType
-                        .GetTypeInfo()
-                        .ImplementedInterfaces
-                        .Where(i => i.GetTypeInfo().IsGenericType
-                                    && i.GenericTypeArguments.Length == 1
-                                    && typeof(AggregateRoot).GetTypeInfo()
-                                        .IsAssignableFrom(i.GenericTypeArguments[0].GetTypeInfo()));
-
-                    foreach (Type serviceType in serviceTypes)
-                    {
-                        Logger.LogDebug(
-                            "Registering scoped authorization service {ServiceType} with implementation {ImplementationType}",
-                            serviceType.Name,
-                            aggregateAuthorizationType.Name);
-                        compositionRoot.Register(
-                            new ServiceDescriptor(
-                                serviceType,
-                                aggregateAuthorizationType,
-                                ServiceLifetime.Scoped));
-                    }
+                    Logger.LogWarning(
+                        "No authorization policies for {AggregateRootTyp} found", aggregateRootType.Name);
+                    return;
                 }
+
+                if (authorizationPolicyTypes.Length > 1)
+                {
+                    throw new InvalidOperationException(
+                        $"Multiple authorization policies found for {aggregateRootType.Name}: " +
+                        $"[{string.Join(", ", authorizationPolicyTypes.Select(t => t.Name))}]");
+                }
+
+                Logger.LogInformation(
+                    "Registering scoped authorization service {ServiceType} with implementation {ImplementationType}",
+                    authorizationPolicyInterfaceType.Name,
+                    authorizationPolicyTypes[0].Name);
+                compositionRoot.Register(ServiceDescriptor.Scoped(authorizationPolicyInterfaceType, authorizationPolicyTypes[0]));
             }
         }
     }
