@@ -1,22 +1,24 @@
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Threading;
 using Backend.Fx.Domain;
+using Backend.Fx.Features.Persistence;
 
 namespace Backend.Fx.Features.Authorization
 {
     /// <summary>
     /// Applies the authorization policy expression to the queryable via decoration
     /// </summary>
-    /// <typeparam name="TAggregateRoot"></typeparam>
-    internal class AuthorizingQueryable<TAggregateRoot> : IAsyncAggregateQueryable<TAggregateRoot> where TAggregateRoot : AggregateRoot
+    internal class AuthorizingQueryable<TAggregateRoot, TId> : IAggregateQueryable<TAggregateRoot, TId>
+        where TAggregateRoot : IAggregateRoot<TId> 
+        where TId : struct, IEquatable<TId>
     {
-        private readonly AuthorizationPolicy<TAggregateRoot> _authorizationPolicy;
-        private readonly IAsyncAggregateQueryable<TAggregateRoot> _aggregateQueryable;
+        private readonly IAuthorizationPolicy<TAggregateRoot, TId> _authorizationPolicy;
+        private readonly IAggregateQueryable<TAggregateRoot, TId> _aggregateQueryable;
 
-        public AuthorizingQueryable(AuthorizationPolicy<TAggregateRoot> authorizationPolicy, IAsyncAggregateQueryable<TAggregateRoot> aggregateQueryable)
+        public AuthorizingQueryable(IAuthorizationPolicy<TAggregateRoot, TId> authorizationPolicy, IAggregateQueryable<TAggregateRoot, TId> aggregateQueryable)
         {
             _authorizationPolicy = authorizationPolicy;
             _aggregateQueryable = aggregateQueryable;
@@ -24,13 +26,32 @@ namespace Backend.Fx.Features.Authorization
 
         public Type ElementType => _aggregateQueryable.ElementType;
 
-        public Expression Expression => Expression.And(_aggregateQueryable.Expression, _authorizationPolicy.HasAccessExpression);
-
-        IAsyncQueryProvider IAsyncQueryable.Provider => _aggregateQueryable.Provider;
-
-        public IAsyncEnumerator<TAggregateRoot> GetAsyncEnumerator(CancellationToken cancellationToken = new CancellationToken())
+        public Expression Expression
         {
-            return _aggregateQueryable.GetAsyncEnumerator(cancellationToken);
+            get
+            {
+                // expression tree manipulation: apply the HasAccessExpression to the basic Queryable Expression using "Where"
+                MethodCallExpression queryableWhereExpression = Expression.Call(
+                    typeof(Queryable),
+                    "Where",
+                    new [] { ElementType },
+                    _aggregateQueryable.Expression, 
+                    Expression.Quote(_authorizationPolicy.HasAccessExpression));
+
+                return queryableWhereExpression;
+            }
+        }
+
+        public IQueryProvider Provider => _aggregateQueryable.Provider;
+
+        public IEnumerator<TAggregateRoot> GetEnumerator()
+        {
+            return _aggregateQueryable.GetEnumerator();
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            return ((IEnumerable)_aggregateQueryable).GetEnumerator();
         }
     }
 }
