@@ -2,27 +2,30 @@ using JetBrains.Annotations;
 
 namespace Backend.Fx.Features.ConfigurationSettings
 {
-    public static class ConfigurationSettingsFeature
+    /// <summary>
+    /// The feature "Configuration Settings" provides a simple abstraction over an arbitrary key/value configuration
+    /// setting store. The default <see cref="SettingSerializerFactory"/> already provides serialization to and from
+    /// string for various configuration setting types, but you can provide your own implementation to extend the
+    /// functionality.
+    /// </summary>
+    /// <typeparam name="TSettingRepository">The abstraction over your key/value store. Instances of this type will
+    /// be injected with a scoped lifetime.</typeparam>
+    [PublicAPI]
+    public class ConfigurationSettingsFeature<TSettingRepository> : Feature
+        where TSettingRepository : class, ISettingRepository
     {
-        /// <summary>
-        /// The feature "Configuration Settings" provides a simple abstraction over an arbitrary key/value configuration
-        /// setting store. The default <see cref="SettingSerializerFactory"/> already provides serialization to and from
-        /// string for various configuration setting types, but you can provide your own implementation to extend the
-        /// functionality.
-        /// </summary>
-        /// <param name="application"></param>
+        private readonly SettingSerializerFactory _settingSerializerFactory;
+
         /// <param name="settingSerializerFactory">The factory that provides serializers. A singleton instance is being held.</param>
-        /// <typeparam name="TSettingRepository">The abstraction over your key/value store. Instances of this type will
-        /// be injected with a scoped lifetime.</typeparam>
-        [PublicAPI]
-        public static void AddConfigurationSettings<TSettingRepository>(
-            this IBackendFxApplication application,
-            SettingSerializerFactory settingSerializerFactory = null)
-            where TSettingRepository : class, ISettingRepository
+        public ConfigurationSettingsFeature(SettingSerializerFactory settingSerializerFactory = null)
         {
-            settingSerializerFactory = settingSerializerFactory ?? new SettingSerializerFactory();
+            _settingSerializerFactory = settingSerializerFactory ?? new SettingSerializerFactory();
+        }
+
+        public override void Enable(IBackendFxApplication application)
+        {
             application.CompositionRoot.RegisterModules(
-                new ConfigurationSettingsModule<TSettingRepository>(settingSerializerFactory));
+                new ConfigurationSettingsModule<TSettingRepository>(_settingSerializerFactory, application.Assemblies));
         }
     }
 }
